20121203
- Added info() calls to try and understand why genus checks are
failing, allowing incompatible types to be assigned to overrides.
 - Found that code only uses genus to know how to reference the
itemname in terms of an agent. There is no check for compatibility with
the selected override.
- Rewrote interface code to check override type and genus, rejecting
incompatible assignment attempts.
 - Works just fine.
 > Fix to be rolled into other functions in the system. FIXME added to
o_addAll - this will need to be re-worked in the same kind of way.
- AddSel button tries to add selected Layout item, not the elements in
the righthand listbox. Documented, but odd, possibly not expected,
behaviour that caused p0010. Opened p0011 for more explicit tracking of
issue and whether this needs to be re-worked.
- Added platform '11' corresponding to Mac64 platform to various
platform checks in the code.
- Added dev flag in the render code to block rendering and instead
permit inspection of the generated scene file and batch file
combination.
- Re-jigged some changes related to override item assignment filtering,
to make them more understandable.
- Minor typo fixes.