20130107
20130107
========

- Misc bug fixes.
- tempDirectory now a global since it's used so much.
- Light override was partially busted (p0016, p0018).
 - Reworked to reliably change/handle lens flare and volumetric
properties.

20130106
========

- Updated graphics to match 'Renewed' differentiation.
- Removed 'Affect OpenGL' light override setting - it has no bearing on
the render.
- Implemented FiberFX pass generation and overrides in Scene Master.
 - Added safety check in scene generation code to query whether we're
altering version 3 data.
 - Changes are very specific in nature due to the unfortunately generic
parameter names being used in this plugin's data.
- Reworked scene generation code due to inconsistent GI parameter
presence in LWS files. Extended the preparation code to review whether
certain inconsistent settings are present. If they aren't, we insert
them into the scene file in the correct location with the corresponding
value to avoid changing behaviour. We then set the value accordingly
later in the generation code.