20130108
- Trying to put a renderer-choice dialog into
createSceneMasterOverride(). Might be borked, so can back this out as
needed.
- Adjusted editSelectedOverride to pick up renderer setting from
override settings.
- Added docs folder containing converted PDFs as .doc files. Will need
to update these nearer release.
- Added misc folder to hold PSD and font stuff that is only useful for
developers.
- Prototype work to support other renderers. Making this modular.
Highly experimental and hardly any code exists right now except for the
native renderer.
 - Major change to refactor code and introduce early value into
settings array for the type6 override to track renderer in use.
 - Will need a partial line read to ensure that Kray is applied before
manipulating. Should warn the user if the data is not found.
- Minor refactoring of sceneGen code to clarify array indexer usage.
- More work on the light override issue (p0016).