20130116
20130116
========

- p0016 *finally* fixed. This revealed a lot of faulty assumptions in
the original code that needed to be adjusted, as well as various issues
in the reworked scene gen system. It's a great relief to have this
fixed, but it made the whole system far more robust in the process.
- Fixed caustics line handling in the scene gen code.
- Removed \n from the volumetric and lens flare settings that are
written out - we only replace lines, not add them in these situations.
We use a different mechanism to make the changes, so \n adds needless
carriage returns.
- Fix to use 'IKInitialState' instead of 'objPathAlignReliableDistLine'
to track end of the standard initial chunk of settings for each block.
- Minor fix to final stage scene gen code.

20130115
========

- Restored missing definition of hvStartLine.
- Working on the final state of the scene generator code, to better
handle the HV stuff and to avoid losing lines in the process.
- p0016 was pretty stubborn, but should be close to resolution.

20130114
========

- Adjusted code to derive number of channels - now uses parse() to
split the line and directly retrieve the number of channels.
- Generic header function added to future-proof header handling.
- Reworked the approach from the original plugin to stripping the
PassEditor reference.
 - No longer breaks line number references, so look-ups of the line
numbers are correct.
 - Color correction settings and related lines are no longer dropped
from the scene. We only strip the PassEditor lines, nothing more,
nothing less.
- Utility functions file added and code moved in to reduce clutter
elsewhere.
- Removed platformVar - replaced with platform()
- Simplifying main scene generation function:
 - Created new function to generate new scene path.
 - Created new functions to hold code to write object, light and camera
data.
 - Created new function to generate SaveRGB path.
- Moved radiosity insertion code to early in the scene generation code
to avoid line reference breakage.
- Optimised some aspects of the scene generation code.