20130119b3
- Autoscan configuration is now supported (p0025). The original code
assumed that configuration files were an either/or thing. You'd either
have the regular EXT file or the Extension Cache. That's not true,
though. You can have both. To that end, we now parse both or either as
we find them. That may lead to some side-effects, but there's no other
mechanism available. Note added to the documentation.