Misc updates
Camera handling with Relativity seemed to be encountering problems,
where the end line was not being retrieved correctly. This is a dirty
fix and is very unlikely to be robust. A better engineered solution is
in-work, but seems to be triggering some flaws in LScript that make it
unsuitable for release at the moment.

Additionally, I’ve added a preference to prevent PassPort deleting
generated batch/bash script files to aid triage of issues. This is
required because both OS X and Windows can close terminal windows and
thus the output is lost.