In-memory changes
This has been sitting on my machine for a while. It’s an experimental
set of changes that avoid using temporary files by making changes
in-memory.
The concern is that LScript has historically been patchy about reliably
keeping data available, and there are potential issues due to the size
of arrays involved.

The aim was to reduce filesystem activity and thus to gain speed, but
it appears that LScript is so slow that the latter is not apparent in
any real world testing. The former might be advantageous for SSDs,
though.